pyproject.toml
[project]
name = "ai-stock-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "crewai[google-genai]>=1.9.3",
    "crewai-tools>=1.9.3",
    "fastapi>=0.135.1",
    "langchain-openai>=0.3.23",
    "litellm>=1.75.3",
    "openai>=1.83.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.1.1",
    "yfinance>=1.2.0",
    "apscheduler>=3.11.2",
    "email-validator>=2.3.0",
    "python-multipart>=0.0.22",
    "fastapi-sso>=0.16.0",
]


[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'"
]
